pyproject.toml
[tool.poetry]
name = "photocard-market"
version = "0.1.0"
description = ""
authors = ["sjinh88 <sjinh88@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2"
uvicorn = "^0.30.1"
djangorestframework = "3.14"
sqlalchemy = "^2.0.31"
psycopg2 = "^2.9.9"
black = "^24.4.2"
isort = "^5.13.2"
pillow = "^10.4.0"
drf-yasg = "^1.21.7"
djangorestframework-simplejwt = "^5.3.1"
mysqlclient = "^2.2.4"
uwsgi = "^2.0.26"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
